Move ExportConfiguration and rename
