+++ hack/generator/pkg/codegen/configuration_test.go
@@ -3,9 +3,10 @@
  * Licensed under the MIT license.
  */
 
-package jsonast
+package codegen
 
 import (
+	"github.com/Azure/k8s-infra/hack/generator/pkg/jsonast"
 	"testing"
 
 	"github.com/Azure/k8s-infra/hack/generator/pkg/astmodel"
@@ -30,7 +31,7 @@ func Test_WithSingleFilter_FiltersExpectedTypes(t *testing.T) {
 	post := post2019
 	student := student2019
 
-	filter := TypeFilter{Action: IncludeType, Version: "2019*"}
+	filter := jsonast.TypeFilter{Action: jsonast.IncludeType, Version: "2019*"}
 	config := NewExportConfiguration(&filter)
 
 	g.Expect(config.ShouldExport(person)).To(Equal(Export))
@@ -45,11 +46,11 @@ func Test_WithMultipleFilters_FiltersExpectedTypes(t *testing.T) {
 	student := student2019
 	address := address2020
 
-	versionFilter := TypeFilter{
-		Action:  IncludeType,
+	versionFilter := jsonast.TypeFilter{
+		Action:  jsonast.IncludeType,
 		Version: "2019*"}
-	nameFilter := TypeFilter{
-		Action: IncludeType,
+	nameFilter := jsonast.TypeFilter{
+		Action: jsonast.IncludeType,
 		Name:   "*ss"}
 	config := NewExportConfiguration(&versionFilter, &nameFilter)
 
@@ -62,11 +63,11 @@ func Test_WithMultipleFilters_FiltersExpectedTypes(t *testing.T) {
 func Test_WithMultipleFilters_GivesPrecedenceToEarlierFilters(t *testing.T) {
 	g := NewGomegaWithT(t)
 
-	alwaysExportPerson := TypeFilter{
-		Action: IncludeType,
+	alwaysExportPerson := jsonast.TypeFilter{
+		Action: jsonast.IncludeType,
 		Name:   "person"}
-	exclude2019 := TypeFilter{
-		Action:  ExcludeType,
+	exclude2019 := jsonast.TypeFilter{
+		Action:  jsonast.ExcludeType,
 		Version: "2019-01-01"}
 	config := NewExportConfiguration(&alwaysExportPerson, &exclude2019)
